Move tab components into routes
