+++ frontend/src/routes/events/OverviewTabs/LeaderboardTab.tsx
